--- src/util/formatTimestamp.ts
@@ -1,15 +1,0 @@
-export const formatTimestamp = (timestamp: string): string => {
-  const date = new Date(timestamp)
-
-  const year = date.getFullYear()
-  const month = String(date.getMonth() + 1).padStart(2, '0')
-  const day = String(date.getDate()).padStart(2, '0')
-
-  const hours = String(date.getHours()).padStart(2, '0')
-  const minutes = String(date.getMinutes()).padStart(2, '0')
-  const seconds = String(date.getSeconds()).padStart(2, '0')
-
-  return `${year}.${month}.${day}. ${hours}:${minutes}:${seconds}`
-}
-
-export default formatTimestamp
